Generalize renderSelection for declarative jobsets
Declarative jobsets require to enter a jobset input while creating a
project, so we need to render the selection based on the new property
system.

This makes inputTypeNames in that we now handle this in renderSelection
directly, because the latter is _only_ used for selecting jobset (and
now project) inputs.

Fixes rendering the jobset input selection while creating projects,
although it doesn't use the property system for the values yet.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>
Cc: @shlevy