Use JSONB field instead project declvalue
In upgrade-51.sql, we have defined a few functions that help in
converting simple values. Because this branch is not yet merged into
master we can safely assume that nobody has deployed this in production
yet.

So we move dropping of these helper functions to upgrade-53.sql instead
so that we can re-use these functions to migrate the "declvalue" in the
projects table as well.

We also apply the build->buildnr fix from upgrade-52.sql as well,
although I find it unlikely that anyone out there using declarative
jobsets is using that combination.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>
Cc: @shlevy