edit-project: Switch to use new property system
It's not complete, because it currently doesn't check the validity of
the properties. But now we can handle the declprops field the same way
as any other jobset input field and thus have more consistency.

I'm using a dummy input field "input-decl-name" here, because we don't
need to rewrite the JavaScript that handles the properties to make room
for properties with a fixed name.

The reason this is there is that if you edit a jobset input you can also
change the name of the input, which doesn't apply for creating/editing
projects.

Another thing that's changed by this is that we now submit the form
values as JSON instead of URL-encoded, which not only makes it easier to
encode but also easier to inspect via the browser's script debugger.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>