properties.js: Provide default if bool is required
If a boolean property is set to false it will default to null, which
removes the field entirely. Unfortunately if the field is required it's
dropped as well, which in turn leads to a validation error because the
property is not defined but required.

This should fix boolean singleton properties in that they now can be set
to false.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>