properties.js: Fix discharging sub-properties
If a sub-property isn't enabled by its parent value (for example if the
parent's property is optional), the check for discharged.value works.

But if the value is a boolean, we get a JS error that we can't access a
boolean value like an object. This is of course true, because we need to
check the discharged value directly (it's not an object but the actual
value).

Signed-off-by: aszlig <aszlig@redmoonstudios.org>