properties.js: Fix initial data for sub-properties
Using {} was a placeholder back then when first writing the
implementation. So far I've solely used sub-properties for secrets so
this didn't get to my attention.

So if values are already present for a given sub-property the values are
used when editing a jobset input or a project, otherwise it will fall
back to {} again.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>