Support builds that can abort themselves
If a build creates the file `/nix-support/aborted` and returns exit
status 0, Hydra will mark the build as aborted, and schedule it for
retry (just like it does for  builds that are manually aborted).
This way, if a build can identify a transient error, it can instruct
Hydra to try the build again.

I believe this could fix the issues #436 and #354.