Consumable system features
Adds support for consumable system features.

Remote build machines can specify a supported or mandatory feature as
consumable with a "feature:n" notation, where n is the quantity of the
feature that the machine provides.  "feature:0" and "feature" are both
interpreted as a non-consumable (infinite) feature.

Derivations can specify how much of a feature they consume with a
similar "feature:n" notation, where n is the quantity of the feature
that the derivation consumes.  "feature" is interpreted as "feature:1"
(i.e., consumes one of the feature), and "feature:0" is interpreted as
still requiring the feature but not consuming any of it.

Closes #585

Signed-off-by: Lisanna Dettwyler <lisanna.dettwyler@gmail.com>