Set notification status to done for dependent builds.
When a build completes with error, the dependent builds are updated
with errors as well, and notifications are sent for all these builds,
but the DB wasn't setting the notification completion indication for
the dependent builds.  One effect of this is that on hydra restart,
notifications would *again* be sent for all dependent build failures,
causing duplicate notifications, many of which were often very old.