Added job/PROJECT/JOBSET/JOB/latest-outpath endpoint.
Returns a redirect to the first output path of the latest successful
build of the specified job. This can be useful when the output
provides HTML documentation (e.g. generated documentation for a build
such as Haskell haddock output).  By providing a "constant" redirect
link, the hydra machine can also serve the documentation it builds via
an href in a static page.