Implement maintainer search
As proposed during the release managers' talk on last NixCon, Hydra
should provide a feature to search for maintainers. I wanted to have
something like this for a long time as well and decided to give it a
try.

This change adds a second search field to the jobset-eval view that can
be used to search for maintainer's github names. Please note that I
intentionally didn't merge this together with the `Search by name`-form
since this form is only used to compare the builds against another eval.

By default Hydra now searches for github handles of maintainers (IMHO this
is more convenient than email addresses), however emails are
used fallbacks as the github handle isn't mandatory for a package
maintainer. The new values are stored in the `Builds` table at the new
column `maintainers_github`.