Draft new structure for maintainers
In order to search for maintainers (emails or github handles), both
values need to be stored. Also, it's probably not desirable to do this
in a string-separated list which is fairly hard to query.

Instead, those values will be stored in a separate table named
`Maintainer` which is associated to the builds-table using a many-to-many[1]
relationship. Please note that I didn't implement a migration for older
builds yet, we probably want to implement an approach similar to the
`hydra-backfill-ids`-implementation[2] here.

Another benefit of this is that if a maintainer with a given email
changes their GitHub-handle, Hydra will recognize this and display
the new handle.

[1] https://metacpan.org/pod/DBIx::Class::Relationship
[2] f69260118bf7e78a4b7830ea58e2f0ba8c85d6b7