Add stale GitHub Action
This GitHub Action will mark and close all issues and PRs that are older
than 180 days.

The list of open PRs and issues is ever growing and the
old issues will probably not ever be looked again. To make it easier for
the maintainers and also to submitters, mark all stale issues/PRs as
such and delete them after 210 days of inactivity.