Temporary hackery to get `maintainers-upgrade-test.nix` running in flake env
It was quite hard to get a test running which also requires an older
version of Hydra. For that, another `nixpkgs` (namely 20.09) was needed.
To make sure that no compilation errors with Nix happen (since neither
Hydra nor Nix with Flakes have *any* releases), Nix is now pinned to a
rev explicitly.