Fix build with the latest Nix master
Fix the build to adjust to some breaking changes in Nix master:

- string2Int now has a slightly different signature
- The printHelp function doesn't exist anymore.

Note that this removes the --help of hydra-eval-jobs.
That doesn't seem too bad given that most hydra commands don't have such a flag, but ideally we should probably migrate it to use the same autogenerated manpages as Nix now uses.