hydra-queue-runner: fix building on machines specified with the `ssh-ng` protocol
Most of my machines in `/etc/nix/machines` are specified as
`ssh-ng://foobar` where `foobar` is defined in e.g. `~/.ssh/config`. The
problem here is that while it's possible to use the remote Nix
installation via SSH, `hydra-queue-runner` refuses to do so since `ssh(1)`
can't resolve hostnames that begin with `ssh-ng://`.

To work around this problem, this prefix will be drop if it exists on a
machine declaration that was chosen for a build step.