api: fix /eval/X/builds requiring 5*len(builds)+1 SQL queries
Prefetch related tables that get fetched as part of serialization to
JSON. Drop the `jobsetevals` field since making it efficient is trickier
(due to not using DBIx native relationships but instead a many_to_many
through a custom table). There is likely a way we could keep it and make
it efficient, but I'm not fluent enough in DBIx to know.