Workaround for fetching git submodules for flakes
If a flake URI contains 'submodules=1' in the query part then any
submodules should be fetched/initialized.

At present, hydra-eval-jobset uses the 'url' field of the flake
metadata as the flake URI.  This URI is missing the 'submodules=1'
query, which is present elsewhere in the metadata.

This PR appends "&submodules=1" if the flake metadata indicates
submodules are used (specifically, if the resolved.submodules field
is true).